Use write_attribute instead of super to set attribute values
In preparation for dropping translated columns.

ActiveRecord attributes that are not bound to database columns are
handled through method_missing so once those columns are dropped, there
will be no named method for super to find.

The handler is a wrapper for write_attribute so we can safely call
that directly ourselves.

https://apidock.com/rails/v4.1.8/ActiveRecord/AttributeMethods/Write/attribute=